Refactor colors/tokens in theme/
