+++ babel.config.js
@@ -13,6 +13,7 @@ module.exports = function (api) {
             '@services': './src/services',
             '@hooks': './src/hooks',
             '@utils': './src/utils',
+            '@theme': './src/theme',
           },
         },
       ],
