+++ src/app/media/gallery/index.tsx
@@ -5,9 +5,9 @@ import { Image, StyleSheet, useWindowDimensions, View } from 'react-native';
 import base64 from 'react-native-base64';
 import { fitContainer, Gallery, GalleryType } from 'react-native-zoom-toolkit';
 import { Post } from '@services/api';
-import { PaletteDark } from '../../../colors';
+import { PaletteDark } from '@theme/colors';
 import Typography from '@components/Typography';
-import { Spacing } from '../../../tokens';
+import { Spacing } from '@theme/tokens';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
 interface GalleryImageProps {
