+++ src/app/media/video/index.tsx
@@ -5,9 +5,9 @@ import * as React from 'react';
 import { Text, TouchableNativeFeedback, View } from 'react-native';
 import base64 from 'react-native-base64';
 import { RedditVideo } from '@services/api';
-import { PaletteDark } from '../../../colors';
+import { PaletteDark } from '@theme/colors';
 import Icons from '@components/Icons';
-import { Spacing } from '../../../tokens';
+import { Spacing } from '@theme/tokens';
 import VideoPlayer from './VideoPlayer';
 import { extractMetaTags } from '@features/video/metadata';
 import Typography from '@components/Typography';
