+++ src/app/settings/videoPlayer/index.tsx
@@ -3,7 +3,7 @@ import useTheme from '@services/theme/useTheme';
 import * as React from 'react';
 import { Switch, View } from 'react-native';
 import { useStore } from '@services/store';
-import { ColorPalette } from '../../../colors';
+import { ColorPalette } from '@theme/colors';
 import Icons, { IconName } from '@components/Icons';
 import Typography from '@components/Typography';
