+++ src/features/home/components/SavedPostsFooter.tsx
@@ -1,9 +1,9 @@
 import { router } from 'expo-router';
 import { TouchableOpacity, View } from 'react-native';
-import { useStore } from '../../../services/store';
-import useTheme from '../../../services/theme/useTheme';
-import Typography from '../../../components/Typography';
-import { Spacing } from '../../../tokens';
+import { useStore } from '@services/store';
+import useTheme from '@services/theme/useTheme';
+import Typography from '@components/Typography';
+import { Spacing } from '@theme/tokens';
 
 const SavedPostsFooter = () => {
   const theme = useTheme();
