+++ src/features/home/components/SearchHeader.tsx
@@ -1,9 +1,9 @@
 import { router } from 'expo-router';
 import { TouchableNativeFeedback, View } from 'react-native';
-import useTheme from '../../../services/theme/useTheme';
-import Icons from '../../../components/Icons';
-import Typography from '../../../components/Typography';
-import { Spacing } from '../../../tokens';
+import useTheme from '@services/theme/useTheme';
+import Icons from '@components/Icons';
+import Typography from '@components/Typography';
+import { Spacing } from '@theme/tokens';
 
 const SearchHeader = () => {
   const theme = useTheme();
