+++ src/features/post/components/PostDetailsHeader.tsx
@@ -3,19 +3,19 @@ import { router } from 'expo-router';
 import { decode } from 'html-entities';
 import React, { useCallback } from 'react';
 import { Pressable, TouchableOpacity, View, useWindowDimensions } from 'react-native';
-import { Post } from '../../../services/api';
-import { ColorPalette } from '../../../colors';
-import CarouselView from '../../../components/CarouselView';
-import Icons from '../../../components/Icons';
-import PostKarmaButton from '../../../components/PostKarmaButton';
-import Typography from '../../../components/Typography';
-import PostFlairChip from '../../../components/PostFlairChip';
-import PostContentPreview from '../../subreddit/feed/components/PostContentPreview';
-import { Spacing } from '../../../tokens';
-import { timeDifference } from '../../../utils/get-time-difference';
+import { Post } from '@services/api';
+import { ColorPalette } from '@theme/colors';
+import CarouselView from '@components/CarouselView';
+import Icons from '@components/Icons';
+import PostKarmaButton from '@components/PostKarmaButton';
+import Typography from '@components/Typography';
+import PostFlairChip from '@components/PostFlairChip';
+import PostContentPreview from '@features/subreddit/feed/components/PostContentPreview';
+import { Spacing } from '@theme/tokens';
+import { timeDifference } from '@utils/get-time-difference';
 import { markdownIt, markdownRenderRules, useMarkdownStyle } from '../utils';
 import PollOption from './PollOption';
-import useGalleryData from '../../../hooks/useGalleryData';
+import useGalleryData from '@hooks/useGalleryData';
 
 function getDisplaySortOrder(forcedSortOrder: string | null, suggestedSort: string | null): string {
   let sort = forcedSortOrder ?? suggestedSort ?? 'best';
