+++ src/features/post/modals/PostDetailsSortOptions.tsx
@@ -1,9 +1,9 @@
 // TODO - move that to /components/ and rename it
 import { TouchableOpacity, ViewStyle } from 'react-native';
-import useTheme from '../../../services/theme/useTheme';
-import Icons, { IconName } from '../../../components/Icons';
-import Typography from '../../../components/Typography';
-import { Spacing } from '../../../tokens';
+import useTheme from '@services/theme/useTheme';
+import Icons, { IconName } from '@components/Icons';
+import Typography from '@components/Typography';
+import { Spacing } from '@theme/tokens';
 import { BottomSheetView } from '@gorhom/bottom-sheet';
 
 interface SortOption<T extends string> {
