+++ src/features/subreddit/feed/modals/ActionSheetOptionRow.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Pressable, View } from 'react-native';
-import useTheme from '../../../../services/theme/useTheme';
-import Icons from '../../../../components/Icons';
-import Typography from '../../../../components/Typography';
+import useTheme from '@services/theme/useTheme';
+import Icons from '@components/Icons';
+import Typography from '@components/Typography';
 
 type Props = {
   title: string;
